Color from popularity
